requirements.txt
Django==5.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
reportlab==4.0.7
django-auditlog==3.0.0
openpyxl==3.1.2
pandas==2.0.3
whitenoise==6.6.0
